getHtmlPieceDescriptors [nfc]: Rename, from `renderMessages`.
The old name was easily confused with, e.g., `renderMessagesAsHtml`,
making it hard to remember (or figure out in the first place) what
each piece of code is responsible for.

We'll also rename `renderMessagesAsHtml`, soon. But before that,
we'll follow up the current (`renderMessages` ->
`getHtmlPieceDescriptors`) rename by propagating it through to code
around its callsites.
