+++ src/message/getHtmlPieceDescriptors.js
