+++ src/message/messageSelectors.js
@@ -11,7 +11,7 @@ import {
 } from '../directSelectors';
 import * as logging from '../utils/logging';
 import { getShownMessagesForNarrow } from '../chat/narrowsSelectors';
-import renderMessages from './renderMessages';
+import getHtmlPieceDescriptors from './getHtmlPieceDescriptors';
 import type { JSONable } from '../utils/jsonable';
 import { ALL_PRIVATE_NARROW_STR } from '../utils/narrow';
 import { NULL_ARRAY } from '../nullObjects';
@@ -64,7 +64,7 @@ export const getPrivateMessages: Selector<Message[]> = createSelector(
 export const getRenderedMessages: Selector<HtmlPieceDescriptor[], Narrow> = createSelector(
   (state, narrow) => narrow,
   getShownMessagesForNarrow,
-  (narrow, messages) => renderMessages(messages, narrow),
+  (narrow, messages) => getHtmlPieceDescriptors(messages, narrow),
 );
 
 export const getFirstUnreadIdInNarrow: Selector<number | null, Narrow> = createSelector(
